+++ src/core/symbols/MidaSymbolParameters.ts
@@ -1,5 +1,4 @@
 import { MidaBrokerAccount } from "#brokers/MidaBrokerAccount";
-import { MidaSymbolSpreadType } from "#symbols/MidaSymbolSpreadType";
 import { MidaSymbolType } from "#symbols/MidaSymbolType";
 
 /** The parameters of the symbol constructor. */
@@ -9,7 +8,6 @@ export type MidaSymbolParameters = {
     description: string;
     type: MidaSymbolType;
     digits: number;
-    spreadType: MidaSymbolSpreadType;
     leverage: number;
     minLots: number;
     maxLots: number;