+++ src/core/symbols/MidaSymbolSpreadType.ts
@@ -1,5 +1,0 @@
-/** The spread type of a symbol. */
-export enum MidaSymbolSpreadType {
-    FIXED = "fixed",
-    FLOATING = "floating",
-}