+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Remove spread type property from symbols.
 * Expose puppeteer page instance in browser tab.
 * Define monthly timeframe type.
 * Create "rightClick" method for browser tabs.